# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Secure Authentication System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  admin: "admin",
  user: "user",
} as const;

export interface RegisterBody {
  /**
   * @minLength 3
   * @maxLength 50
   */
  username: string;
  /** @minLength 6 */
  password: string;
  role?: RegisterBodyRole;
}

export interface LoginBody {
  username: string;
  password: string;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  admin: "admin",
  user: "user",
} as const;

export interface UserProfile {
  id: number;
  username: string;
  role: UserProfileRole;
  createdAt: string;
}

export interface AuthResponse {
  user: UserProfile;
  message: string;
}

export type AdminUserViewRole =
  (typeof AdminUserViewRole)[keyof typeof AdminUserViewRole];

export const AdminUserViewRole = {
  admin: "admin",
  user: "user",
} as const;

export interface AdminUserView {
  id: number;
  username: string;
  role: AdminUserViewRole;
  failedAttempts: number;
  isLocked: boolean;
  createdAt: string;
}

export interface AuditLog {
  id: number;
  userId?: number | null;
  username?: string | null;
  action: string;
  ipAddress?: string | null;
  details?: string | null;
  createdAt: string;
}

export interface SystemStats {
  totalUsers: number;
  lockedAccounts: number;
  adminCount: number;
  userCount: number;
  totalLoginAttempts: number;
  failedLoginAttempts: number;
  recentActivity: AuditLog[];
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginErrorResponse {
  error: string;
  locked?: boolean;
  attemptsRemaining?: number;
}

export type AdminGetLogsParams = {
  limit?: number;
  action?: string;
};