# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Secure Authentication System API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new user
 */
export const registerBodyUsernameMin = 3;
export const registerBodyUsernameMax = 50;

export const registerBodyPasswordMin = 6;

export const registerBodyRoleDefault = `user`;

export const RegisterBody = zod.object({
  username: zod
    .string()
    .min(registerBodyUsernameMin)
    .max(registerBodyUsernameMax),
  password: zod.string().min(registerBodyPasswordMin),
  role: zod.enum(["admin", "user"]).default(registerBodyRoleDefault),
});

/**
 * @summary Login with credentials
 */
export const LoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  user: zod.object({
    id: zod.number(),
    username: zod.string(),
    role: zod.enum(["admin", "user"]),
    createdAt: zod.coerce.date(),
  }),
  message: zod.string(),
});

/**
 * @summary Logout current user
 */
export const LogoutResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Get current authenticated user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  role: zod.enum(["admin", "user"]),
  createdAt: zod.coerce.date(),
});

/**
 * @summary List all users (admin only)
 */
export const AdminListUsersResponseItem = zod.object({
  id: zod.number(),
  username: zod.string(),
  role: zod.enum(["admin", "user"]),
  failedAttempts: zod.number(),
  isLocked: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const AdminListUsersResponse = zod.array(AdminListUsersResponseItem);

/**
 * @summary Unlock a locked user account (admin only)
 */
export const AdminUnlockUserParams = zod.object({
  id: zod.coerce.number(),
});

export const AdminUnlockUserResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Get audit logs (admin only)
 */
export const adminGetLogsQueryLimitDefault = 50;

export const AdminGetLogsQueryParams = zod.object({
  limit: zod.coerce.number().default(adminGetLogsQueryLimitDefault),
  action: zod.coerce.string().optional(),
});

export const AdminGetLogsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number().nullish(),
  username: zod.string().nullish(),
  action: zod.string(),
  ipAddress: zod.string().nullish(),
  details: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});
export const AdminGetLogsResponse = zod.array(AdminGetLogsResponseItem);

/**
 * @summary Get system statistics (admin only)
 */
export const AdminGetStatsResponse = zod.object({
  totalUsers: zod.number(),
  lockedAccounts: zod.number(),
  adminCount: zod.number(),
  userCount: zod.number(),
  totalLoginAttempts: zod.number(),
  failedLoginAttempts: zod.number(),
  recentActivity: zod.array(
    zod.object({
      id: zod.number(),
      userId: zod.number().nullish(),
      username: zod.string().nullish(),
      action: zod.string(),
      ipAddress: zod.string().nullish(),
      details: zod.string().nullish(),
      createdAt: zod.coerce.date(),
    }),
  ),
});